refactor(backup): split backup_page; finalise test relative-import paths
backup/widgets/backup_body.dart hosts BackupBody (the two-button
column the FutureBuilder renders once the backup-file path resolves).
Page file is now state + delegate-to-service only.

Final import-path fixups for tests that moved deeper in the previous
commits (calendar, backup, app smoke) — pure mechanical change to
walk up the right number of directories to test/helpers/ and
test/test_helper.dart.

Suite tally: 103 tests pass, ~30 widgets now have per-file lcov
coverage entries that didn't exist when they lived inline. Pages
themselves became thin orchestrators (60-200 lines each).
